third_party/WebKit/LayoutTests/fast/notifications/notifications-permission-values.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <body>
        <p>
            Tests that the return values of Notification.permission match
            the permission request status.
        </p>
        <div id="console"></div>
        <script>
            if (!window.testRunner) {
                testFailed("This test depends on window.testRunner to toggle permissions.");
            } else {
                testRunner.clearWebNotificationPermissions();

                // The "default" state indicates that no permission request has
                // been answered yet.
                shouldBeEqualToString("Notification.permission", "default");

                testRunner.grantWebNotificationPermission("file://", false);
                shouldBeEqualToString("Notification.permission", "denied");

                testRunner.grantWebNotificationPermission("file://", true);
                shouldBeEqualToString("Notification.permission", "granted");
            }
        </script>
    </body>
</html>
